+++ src/Fundme.sol
@@ -18,7 +18,6 @@ contract FundMe {
     uint256 public constant MINIMUM_USD = 5 * 10 ** 18;
     AggregatorV3Interface private s_priceFeed;
 
-
     constructor(address priceFeed) {
         i_owner = msg.sender;
         s_priceFeed = AggregatorV3Interface(priceFeed);
@@ -40,21 +39,22 @@ contract FundMe {
         if (msg.sender != i_owner) revert NotOwner();
         _;
     }
-    function cheaperWithdraw() public onlyOwner{
-        uint fundersLength=s_funders.length;
-       for(uint funderIndex=0;funderIndex<fundersLength;funderIndex++){
-           address funder = s_funders[funderIndex];
-           s_addressToAmountFunded[funder] = 0;
-       }
-       s_funders = new address[](0);
-       (bool callSuccess,) = payable(msg.sender).call{value: address(this).balance}("");
+
+    function cheaperWithdraw() public onlyOwner {
+        uint256 fundersLength = s_funders.length;
+        for (uint256 funderIndex = 0; funderIndex < fundersLength; funderIndex++) {
+            address funder = s_funders[funderIndex];
+            s_addressToAmountFunded[funder] = 0;
+        }
+        s_funders = new address[](0);
+        (bool callSuccess,) = payable(msg.sender).call{value: address(this).balance}("");
         require(callSuccess, "Call failed");
-   }
+    }
 
     function withdraw() public onlyOwner {
         for (uint256 funderIndex = 0; funderIndex < s_funders.length; funderIndex++) {
             address funder = s_funders[funderIndex];
-           s_addressToAmountFunded[funder] = 0;
+            s_addressToAmountFunded[funder] = 0;
         }
         s_funders = new address[](0);
         // // transfer
@@ -87,16 +87,18 @@ contract FundMe {
     receive() external payable {
         fund();
     }
-    function getAddressToAmountFunded(address fundingAddress) external view returns(uint){
+
+    function getAddressToAmountFunded(address fundingAddress) external view returns (uint256) {
         return s_addressToAmountFunded[fundingAddress];
     }
-    function getFunder(uint index) external view returns(address){
+
+    function getFunder(uint256 index) external view returns (address) {
         return s_funders[index];
     }
-    function getOwner() external view returns(address){
+
+    function getOwner() external view returns (address) {
         return i_owner;
     }
-
 }
 
 // Concepts we didn't cover yet (will cover in later sections)
@@ -107,4 +109,3 @@ contract FundMe {
 // 5. abi.encode / decode
 // 6. Hash with keccak256
 // 7. Yul / Assembly
-
