+++ test/Integration/Interaction.t.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.18;
+
 import {Test, console} from "forge-std/Test.sol";
 import {FundMe} from "../../src/Fundme.sol";
 import {DeployFundMe} from "../../script/DeployFundMe.s.sol";
@@ -37,9 +38,9 @@ contract InteractionsTest is Test {
     FundMe public fundMe;
     DeployFundMe deployFundMe;
     address USER = makeAddr("user");
-    uint constant SEND_VALUE = 0.1 ether;
-    uint constant STARTING_BALANCE = 10 ether;
-    uint constant GAS_PRICE = 1;
+    uint256 constant SEND_VALUE = 0.1 ether;
+    uint256 constant STARTING_BALANCE = 10 ether;
+    uint256 constant GAS_PRICE = 1;
 
     function setUp() external {
         deployFundMe = new DeployFundMe();
